# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api-rbac_internal_model.CheckPermissionRequest:
    properties:
      action:
        example: delete
        maxLength: 64
        minLength: 1
        type: string
      resource:
        example: user
        maxLength: 64
        minLength: 1
        type: string
    required:
    - action
    - resource
    type: object
  api-rbac_internal_model.CreatePermissionRequest:
    properties:
      action:
        example: delete
        maxLength: 64
        minLength: 2
        type: string
      description:
        example: 允许删除其他用户账号
        maxLength: 255
        type: string
      name:
        example: 删除用户
        maxLength: 64
        minLength: 2
        type: string
      resource:
        example: user
        maxLength: 64
        minLength: 2
        type: string
    required:
    - action
    - name
    - resource
    type: object
...
  api-rbac_internal_model.LoginRequest:
    properties:
      account:
        example: admin
        maxLength: 128
        minLength: 2
        type: string
      password:
        example: admin123
        maxLength: 128
        type: string
    required:
    - account
    - password
    type: object
  api-rbac_internal_model.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
      updated_at:
        type: string
    type: object
...
  api-rbac_internal_model.UpdatePermissionRequest:
    properties:
      action:
        example: delete
        maxLength: 64
        minLength: 2
        type: string
      description:
        example: 允许删除任意用户账号
        maxLength: 255
        type: string
      name:
        example: 删除任意用户
        maxLength: 64
        minLength: 2
        type: string
      resource:
        example: user
        maxLength: 64
        minLength: 2
        type: string
    required:
    - action
    - name
    - resource
    type: object
...
  api-rbac_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  api-rbac_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: 通用的 RBAC 权限管理微服务，支持用户/角色/权限的 CRUD 及绑定关系管理。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RBAC 权限管理系统 API
  version: "1.0"
paths:
  /auth/check:
    post:
      consumes:
      - application/json
      description: 检查当前用户是否拥有对指定资源的指定操作权限。业务系统在执行敏感操作前调用此接口做鉴权。
      parameters:
      - description: 权限检查参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.CheckPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 有权限
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  properties:
                    allowed:
                      type: boolean
                  type: object
              type: object
        "403":
          description: 无权限
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  properties:
                    allowed:
                      type: boolean
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 检查权限
      tags:
      - 认证
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名或邮箱登录，返回 JWT Token
      parameters:
      - description: 登录参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  properties:
                    token:
                      type: string
                    user_id:
                      type: integer
                    username:
                      type: string
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "401":
          description: 密码错误或用户禁用
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      description: 无状态 JWT，客户端丢弃 Token 即完成登出
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证
  /auth/verify:
    post:
      description: 验证 JWT Token 是否有效，返回用户基本信息。业务系统在收到请求时可调用此接口确认 Token 有效性。
      produces:
      - application/json
      responses:
        "200":
          description: Token有效
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  properties:
                    user_id:
                      type: integer
                    username:
                      type: string
                  type: object
              type: object
        "401":
          description: Token无效或已过期
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 验证Token
      tags:
      - 认证
  /permissions:
    get:
      description: 分页查询权限列表，支持关键词搜索
      parameters:
      - default: 1
        description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        example: 10
        in: query
        name: page_size
        type: integer
      - description: 搜索关键词(权限名/资源/操作)
        example: 删除
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_pkg_response.PageData'
              type: object
      security:
      - BearerAuth: []
      summary: 权限列表
      tags:
      - 权限管理
    post:
      consumes:
      - application/json
      description: 新增一个权限定义
      parameters:
      - description: 创建权限参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_internal_model.Permission'
              type: object
        "400":
          description: 参数错误或权限名已存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建权限
      tags:
      - 权限管理
  /permissions/{id}:
    delete:
      description: 根据 ID 删除一个权限
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 权限不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除权限
      tags:
      - 权限管理
    get:
      description: 根据 ID 获取权限信息
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/api-rbac_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api-rbac_internal_model.Permission'
              type: object
        "404":
          description: 权限不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取权限详情
      tags:
      - 权限管理
    put:
      consumes:
      - application/json
      description: 更新权限的名称、资源和操作
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api-rbac_internal_model.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
        "404":
          description: 权限不存在
          schema:
            $ref: '#/definitions/api-rbac_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新权限
      tags:
      - 权限管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token} 格式的 JWT Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"